pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roborock-cloud-cli"
version = "0.4.0"
description = "OpenClaw-compatible Roborock CLI and agent skill backend"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "MOZARTINOS" }]
keywords = [
  "openclaw",
  "openclaw-skill",
  "agent",
  "automation",
  "roborock",
  "vacuum",
  "smart-home",
  "home-automation",
  "cli",
  "mqtt",
  "telegram-bot",
  "webrtc",
  "camera",
  "adb",
  "rooms",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Home Automation",
]
dependencies = [
  "python-roborock>=4.17.0",
  "aiohttp>=3.9.0",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
camera = ["aiortc>=1.9.0", "av>=12.0.0"]
all = ["python-telegram-bot>=21.0", "aiortc>=1.9.0", "av>=12.0.0"]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.urls]
Homepage = "https://github.com/MOZARTINOS/openclaw-roborock-cli"
Issues = "https://github.com/MOZARTINOS/openclaw-roborock-cli/issues"
Documentation = "https://github.com/MOZARTINOS/openclaw-roborock-cli#readme"

[project.scripts]
roborock-cli = "roborock_cli.cli:main"

[tool.setuptools.packages.find]
include = ["roborock_cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
